- keybind listening thread implimented, keybinds now work - user can now speak while playing sounds
NOTE
at this point in development, the biggest concern of mine is polishing the UI and user feedback system, presently it is not very pretty, or evident what is happening and when.

Next i will be implimenting the settings and add sound pages, but i will just make some themes for now as my brain is fried.